requirements.txt
milvus
pdfminer-six
sentence-transformers
ollama
streamlit
pymilvus